script fixes,
