+++ l1-contracts/scripts/register-hyperchain.ts
@@ -8,7 +8,7 @@ import * as fs from "fs";
 import * as path from "path";
 import { Deployer } from "../src.ts/deploy";
 import { GAS_MULTIPLIER, web3Provider } from "./utils";
-import { ADDRESS_ONE } from "../src.ts/utils";
+import { ADDRESS_ONE } from "../src.ts/constants";
 import { getTokens } from "../src.ts/deploy-token";
 
 const ETH_TOKEN_ADDRESS = ADDRESS_ONE;
