+++ l1-contracts/scripts/setup-legacy-bridge-era.ts
@@ -13,7 +13,8 @@ import { Deployer } from "../src.ts/deploy";
 import { formatUnits, parseUnits, Interface } from "ethers/lib/utils";
 import { web3Provider, GAS_MULTIPLIER } from "./utils";
 import { deployedAddressesFromEnv } from "../src.ts/deploy-utils";
-import { ethTestConfig, getAddressFromEnv } from "../src.ts/utils";
+import { ethTestConfig } from "../src.ts/constants";
+import { getAddressFromEnv } from "../src.ts/utils";
 import { hashL2Bytecode } from "../../l2-contracts/src/utils";
 import { Provider } from "zksync-ethers";
 import beaconProxy = require("../../l2-contracts/artifacts-zk/@openzeppelin/contracts/proxy/beacon/BeaconProxy.sol/BeaconProxy.json");
