+++ l1-contracts/src.ts/deploy-test-process.ts
@@ -20,16 +20,9 @@ import {
 import { deployTokens, getTokens } from "./deploy-token";
 
 import { SYSTEM_CONFIG } from "../scripts/utils";
-import {
-  testConfigPath,
-  getNumberFromEnv,
-  getHashFromEnv,
-  PubdataPricingMode,
-  ADDRESS_ONE,
-  EMPTY_STRING_KECCAK,
-  isCurrentNetworkLocal,
-  ETH_ADDRESS_IN_CONTRACTS,
-} from "./utils";
+import { getNumberFromEnv, getHashFromEnv, PubdataPricingMode, isCurrentNetworkLocal } from "./utils";
+import { testConfigPath, ADDRESS_ONE, EMPTY_STRING_KECCAK, ETH_ADDRESS_IN_CONTRACTS } from "../src.ts/constants";
+
 import { diamondCut, getCurrentFacetCutsForAdd, facetCut, Action } from "./diamondCut";
 import { CONTRACTS_GENESIS_PROTOCOL_VERSION } from "../test/unit_tests/utils";
