+++ l1-contracts/test/unit_tests/custom_base_token.spec.ts
@@ -11,7 +11,7 @@ import { IL1AssetRouterFactory } from "../../typechain/IL1AssetRouterFactory";
 
 import { getTokens } from "../../src.ts/deploy-token";
 import type { Deployer } from "../../src.ts/deploy";
-import { ethTestConfig } from "../../src.ts/utils";
+import { ethTestConfig } from "../../src.ts/constants";
 import { initialTestnetDeploymentProcess } from "../../src.ts/deploy-test-process";
 
 import { getCallRevertReason, REQUIRED_L2_GAS_PRICE_PER_PUBDATA } from "./utils";
