+++ l1-contracts/test/unit_tests/initial_deployment_test.spec.ts
@@ -7,7 +7,7 @@ import type { Bridgehub, StateTransitionManager } from "../../typechain";
 import { BridgehubFactory, StateTransitionManagerFactory } from "../../typechain";
 
 import { initialTestnetDeploymentProcess } from "../../src.ts/deploy-test-process";
-import { ethTestConfig } from "../../src.ts/utils";
+import { ethTestConfig } from "../../src.ts/constants";
 
 import type { Deployer } from "../../src.ts/deploy";
