+++ l1-contracts/test/unit_tests/synclayer.spec.ts
@@ -11,14 +11,15 @@ import {
   defaultDeployerForTests,
   registerHyperchainWithBridgeRegistration,
 } from "../../src.ts/deploy-test-process";
+import { priorityTxMaxGasLimit } from "../../src.ts/utils";
 import {
   ethTestConfig,
   DIAMOND_CUT_DATA_ABI_STRING,
   HYPERCHAIN_COMMITMENT_ABI_STRING,
   ADDRESS_ONE,
   REQUIRED_L2_GAS_PRICE_PER_PUBDATA,
-  priorityTxMaxGasLimit,
-} from "../../src.ts/utils";
+} from "../../src.ts/constants";
+
 import { SYSTEM_CONFIG } from "../../scripts/utils";
 
 import type { Deployer } from "../../src.ts/deploy";
